+++ src/view/amount-films.js
@@ -1,4 +1,5 @@
 //Количество фильмов а футере
 export const createNumberFilmTemplate = () => (
-  `<p>130 291 movies inside</p>`
+  `<p>130 291 movies inside
+  </p>`
 );
