+++ src/view/show-more.js
@@ -1,0 +1,5 @@
+//Кнопка показать больше фильмов
+export const createButtonShowTemplate = () => (
+  `<button class="films-list__show-more">Show more
+  </button>`
+);
